feat(ui): 组网统一节点模型——TS 融入统一卡 + WARP 单例 + 退役 TailscaleConnectionCard
- server-page 停止滤出 TS + 删 TailscaleConnectionCard; TS 作为统一 ServerCard(登录态 badge/current ring/ⓘ), 点卡=选出口
- 新增 mesh-tailscale-control(5 态连接生命周期, #174/#254 handler 逐字迁移: 取消乐观清/打开登录页回落/loginActive 门控/出口警示 auto-open 全保住)
- 重写 mesh-access-entry: WG 恒新增/TS 已接入管理/WARP 单例(findWarpNode+warpSlotTaken 守卫, 注销走 WarpDeregisterQueue 防孤儿计费)
- 补登录态→卡角标映射 + WARP 单例守卫单测; 5 语言 i18n parity; deriveTsCardState/emitTailscaleStatus/mesh-login-fallback 未触碰
