+++ src/renderer/i18n/locales/en-US.json
@@ -1170,6 +1170,15 @@
     "meshAccessAddWarp": "Add Cloudflare WARP",
     "meshAccessAddWarpDesc": "One click registers an anonymous WARP device and adds it as a WireGuard node.",
     "meshAccessNameFirst": "Enter a node name first",
+    "meshManage": "Manage",
+    "meshAddWireguard": "Add WireGuard",
+    "meshWarpJoined": "WARP connected",
+    "meshWarpReRegister": "Re-register",
+    "meshWarpDeregister": "Remove",
+    "meshWarpReRegisterTitle": "Re-register WARP?",
+    "meshWarpReRegisterDesc": "Deregisters the current WARP device, then registers a fresh anonymous device to replace it.",
+    "meshWarpDeregisterTitle": "Remove WARP?",
+    "meshWarpDeregisterDesc": "Removes this WARP node and deregisters the remote anonymous device in the background. You can rejoin anytime.",
     "tsReauth": "Re-login · switch account",
     "tsLogoutRestartHint": "Node is running; restart the proxy to take effect",
     "tsLoggedInNoKeyHint": "Logged in — no auth key needed",
