+++ src/renderer/i18n/locales/fa.json
@@ -1170,6 +1170,15 @@
     "meshAccessAddWarp": "افزودن Cloudflare WARP",
     "meshAccessAddWarpDesc": "با یک کلیک یک دستگاه WARP ناشناس ثبت می‌شود و به‌عنوان گره WireGuard اضافه می‌شود.",
     "meshAccessNameFirst": "ابتدا یک نام گره وارد کنید",
+    "meshManage": "مدیریت",
+    "meshAddWireguard": "افزودن WireGuard",
+    "meshWarpJoined": "WARP متصل شد",
+    "meshWarpReRegister": "ثبت‌نام مجدد",
+    "meshWarpDeregister": "حذف",
+    "meshWarpReRegisterTitle": "ثبت‌نام مجدد WARP؟",
+    "meshWarpReRegisterDesc": "ابتدا ثبت دستگاه فعلی WARP لغو می‌شود، سپس یک دستگاه ناشناس جدید جایگزین آن ثبت می‌شود.",
+    "meshWarpDeregisterTitle": "حذف WARP؟",
+    "meshWarpDeregisterDesc": "این گره WARP را حذف کرده و در پس‌زمینه ثبت دستگاه ناشناس راه دور را لغو می‌کند. هر زمان می‌توانید دوباره متصل شوید.",
     "tsReauth": "ورود مجدد · تعویض حساب",
     "tsLogoutRestartHint": "گره در حال اجراست؛ برای اعمال، پروکسی را دوباره راه‌اندازی کنید",
     "tsLoggedInNoKeyHint": "وارد شده — به کلید احراز هویت نیازی نیست",
