+++ src/renderer/i18n/locales/ru.json
@@ -1170,6 +1170,15 @@
     "meshAccessAddWarp": "Добавить Cloudflare WARP",
     "meshAccessAddWarpDesc": "Один клик регистрирует анонимное устройство WARP и добавляет его как узел WireGuard.",
     "meshAccessNameFirst": "Сначала введите имя узла",
+    "meshManage": "Управление",
+    "meshAddWireguard": "Добавить WireGuard",
+    "meshWarpJoined": "WARP подключён",
+    "meshWarpReRegister": "Перерегистрировать",
+    "meshWarpDeregister": "Удалить",
+    "meshWarpReRegisterTitle": "Перерегистрировать WARP?",
+    "meshWarpReRegisterDesc": "Сначала будет отменена регистрация текущего устройства WARP, затем зарегистрировано новое анонимное устройство взамен.",
+    "meshWarpDeregisterTitle": "Удалить WARP?",
+    "meshWarpDeregisterDesc": "Удаляет этот узел WARP и в фоне отменяет регистрацию удалённого анонимного устройства. Можно переподключиться в любой момент.",
     "tsReauth": "Повторный вход · сменить аккаунт",
     "tsLogoutRestartHint": "Узел запущен; перезапустите прокси, чтобы применить",
     "tsLoggedInNoKeyHint": "Вход выполнен — ключ авторизации не нужен",
