+++ src/renderer/i18n/locales/zh-CN.json
@@ -1170,6 +1170,15 @@
     "meshAccessAddWarp": "添加 Cloudflare WARP",
     "meshAccessAddWarpDesc": "一键注册匿名 WARP 设备并加入为 WireGuard 节点。",
     "meshAccessNameFirst": "请先填写节点名称",
+    "meshManage": "管理",
+    "meshAddWireguard": "新增 WireGuard",
+    "meshWarpJoined": "已接入 WARP",
+    "meshWarpReRegister": "重新注册",
+    "meshWarpDeregister": "注销",
+    "meshWarpReRegisterTitle": "重新注册 WARP？",
+    "meshWarpReRegisterDesc": "将先注销当前 WARP 设备，再注册一台全新的匿名设备替换它。",
+    "meshWarpDeregisterTitle": "注销 WARP？",
+    "meshWarpDeregisterDesc": "将移除此 WARP 节点并在后台注销远端匿名设备，可随时重新接入。",
     "tsReauth": "重新登录 / 切换账号",
     "tsLogoutRestartHint": "该节点运行中，退出后需重启代理生效",
     "tsLoggedInNoKeyHint": "已登录，无需 Auth Key",
