+++ src/renderer/i18n/locales/zh-TW.json
@@ -1170,6 +1170,15 @@
     "meshAccessAddWarp": "新增 Cloudflare WARP",
     "meshAccessAddWarpDesc": "一鍵註冊匿名 WARP 裝置並加入為 WireGuard 節點。",
     "meshAccessNameFirst": "請先填寫節點名稱",
+    "meshManage": "管理",
+    "meshAddWireguard": "新增 WireGuard",
+    "meshWarpJoined": "已接入 WARP",
+    "meshWarpReRegister": "重新註冊",
+    "meshWarpDeregister": "註銷",
+    "meshWarpReRegisterTitle": "重新註冊 WARP？",
+    "meshWarpReRegisterDesc": "將先註銷目前的 WARP 裝置，再註冊一台全新的匿名裝置替換它。",
+    "meshWarpDeregisterTitle": "註銷 WARP？",
+    "meshWarpDeregisterDesc": "將移除此 WARP 節點並在背景註銷遠端匿名裝置，可隨時重新接入。",
     "tsReauth": "重新登錄 / 切換帳號",
     "tsLogoutRestartHint": "該節點運行中，退出後需重啟代理生效",
     "tsLoggedInNoKeyHint": "已登錄，無需 Auth Key",
